+++ main.js
@@ -16,43 +16,58 @@ const https = require('https');
 const readline = require('readline');
 const tough = require('tough-cookie');
 
+// Spec holder class. Please see comments of member variables for details.
+class PhishingServer {
+  constructor(__domain, __entry_path = '/', __login_page = '/home/login', __login_inbox = '/login/dologin') {
+
+    // Target C&C server domain name.
+    this.target_domain = __domain;
+
+    // Remote path for server cookie init.
+    this.entry_path = __entry_path;
+
+    // Remote path of the phishing page.
+    this.phishing_page_path = __login_page;
+
+    // Remote path of the phishing data receiver.
+    this.phishing_target_path = __login_inbox;
+  }
+}
+
 /*
  * Settings are here c:
  * Modify them on your purpose.
  */
 
-// Target C&C server domain name.
-// Some malicious servers:
-//  [0] => steamcomrninuty.ru (seized)
-//  [1] => steamcommunity.link
-//  [2] => steamcommnunmity.com
-//  [3] => discorcl.link
-let spam_target = 'steamcommunity.link';
-
-// Remote path for server cookie init.
-//  [*] => /
-//  [2] => /id/zxayonax
-let entry_path = '/';
-
-// Remote path of the phishing page.
-//  [*] => /home/login
-//  [3] => /BT7D8pw3BYODgyTqxB8BzIbCyN9FSD30MirWfrMbd4a5a8d2fYvUHMxgt
-let phishing_page_path = '/home/login';
-
-// Remote path of the phishing data receiver.
-let phishing_target_path = '/login/dologin';
+// List of malicious servers.
+let spam_targets = [
+  // [seized operation] new PhishingServer('steamcomrninuty.ru'),
+  // [seized operation] new PhishingServer('steamcommnunmity.com', '/id/zxayonax'),
+  new PhishingServer('steamcommunity.link'),
+  new PhishingServer('discorcl.link'),
+  new PhishingServer('nitro-discord.ru', '/airdrop'),
+];
 
 // Debug mode: only one attempt and verbose output. Used for testing if the spam works.
 let debug_mode = false;
 
 // Number of connection procedures in parallel.
-let parallel_count = 1;
+let parallel_count = 4;
 
 // Random TCP error count threshold for each parallel member before abortion.
 let error_count_threshold = 8;
 
 // Upper limit of the number of spam messages.
 let spam_limit = 1000;
+
+// Determine if the process should continue after receiving HTTP 429 (Too Many Requests).
+//  If this is set to a positive value (in seconds), the process will sleep for the spicified time before restart spamming;
+//  Else, the process terminates after receiving HTTP 429.
+let sleep_when_429 = 300;
+
+// When sleep_when_429 is set and all servers are returning HTTP 429 errors, the target pool may be empty for a while.
+//  The members will constantly check if the target pool has become available again with this interval in milliseconds.
+let revive_interval = 250;
 
 // User-Agent for faking browser connection.
 let useragent = 'Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.79 Safari/537.36';
@@ -62,6 +77,16 @@ let useragent = 'Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KH
 /*
  * Runtime variables and functions below. Do not edit unless needed.
  */
+
+// Target pool. Used to deal with sleeping functionality for HTTP 429.
+let target_pool = new Map();
+for(let i = 0 ; i < spam_targets.length ; i++) {
+  spam_targets[i].index = i;
+  target_pool.set(spam_targets[i].target_domain, spam_targets[i]);
+}
+
+// Sleep timeout handler objects.
+let timeout_handles_targets = new Array(spam_targets.length);
 
 // Readline console object.
 let reli = null;
@@ -73,7 +98,7 @@ let stopping = false;
 let error_count = new Array(parallel_count);
 
 // Pool for fake password.
-let mass_char_sets = Buffer.from('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789');
+let mass_char_sets = Buffer.from('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_____');
 
 // Tough-cookie init.
 let Cookie = tough.Cookie;
@@ -96,11 +121,21 @@ let dispatched_spams = 0;
 // Incremental message ID cache for each member.
 let spam_sess_idx = new Array(parallel_count);
 
+// Sleep timeout handler objects.
+let timeout_handles_member = new Array(parallel_count);
+
+// Cached targets for each session.
+let spam_targets_cache = new Array(parallel_count);
+
+function randomU32() {
+  return Math.floor(Math.random() * 0x100000000);
+}
+
 // Generate a random string as fake username or password.
 function generate_random_string() {
   let ret = '';
   for (let i = 0; i < 2; i++) {
-    let int01 = Math.floor(Math.random() * 0x100000000);
+    let int01 = randomU32();
     while (int01 > 0) {
       ret = `${String.fromCharCode(mass_char_sets[int01 % mass_char_sets.length])}${ret}`;
       int01 = Math.floor(int01 / mass_char_sets.length);
@@ -120,9 +155,14 @@ function kill_spam_member()
   spam_members_alive--;
 
   // No member is alive, close the console to avoid zombifying
-  if(spam_members_alive <= 0)
-  {
+  if(spam_members_alive <= 0) {
     reli.close();
+    for (let i = 0 ; i < spam_targets.length ; i++) {
+      if (timeout_handles_targets[i]) {
+        clearTimeout(timeout_handles_targets[i]);
+        timeout_handles_targets[i] = null;
+      }
+    }
   }
 }
 
@@ -130,6 +170,7 @@ function kill_spam_member()
 // Too bad we cannot execute them in public and drain their dirty blood into sewers alive.
 function do_spam(idx) {
   if (stopping) {
+    kill_spam_member();
     return;
   }
 
@@ -141,16 +182,33 @@ function do_spam(idx) {
         kill_spam_member();
         return;
       }
+
+      // Check targets availibility
+      if (target_pool.size > 0) {
+        timeout_handles_member[idx] = null;
+        spam_targets_cache[idx] = target_pool.get(Array.from(target_pool.keys())[randomU32() % target_pool.size]);
+      } else {
+        if (sleep_when_429 > 0) {
+          timeout_handles_member[idx] = setTimeout(do_spam, revive_interval, idx);
+        } else {
+          stopping = true;
+          kill_spam_member();
+        }
+        return;
+      }
+
       spam_sess_idx[idx] = dispatched_spams;
       dispatched_spams++;
 
-      let req1 = https.request(`https://${spam_target}${entry_path}`, {
+      let req1 = https.request(`https://${spam_targets_cache[idx].target_domain}${spam_targets_cache[idx].entry_path}`, {
         method: 'GET',
         headers: {
           'User-Agent': useragent
         }
       });
       req1.end();
+
+      console.log('[%s:%s] Target: %s', `0${idx}`.slice(-2), `000000${spam_sess_idx[idx]}`.slice(-7), spam_targets_cache[idx].target_domain);
 
       req1.on('response', function (resp1) {
         let buf1 = Buffer.allocUnsafe(0);
@@ -188,6 +246,8 @@ function do_spam(idx) {
           // Proceed to next stage if not stopping.
           if (!stopping) {
             setImmediate(do_spam, idx);
+          } else {
+            kill_spam_member();
           }
         });
       });
@@ -207,7 +267,7 @@ function do_spam(idx) {
 
     // Intermediate stage, where _tdg is present but not lumen_session.
     case 1: {
-      let req1 = https.request(`https://${spam_target}${entry_path}`, {
+      let req1 = https.request(`https://${spam_targets_cache[idx].target_domain}${spam_targets_cache[idx].entry_path}`, {
         method: 'GET',
         headers: {
           'User-Agent': useragent,
@@ -246,6 +306,8 @@ function do_spam(idx) {
           // Proceed to next stage if not stopping.
           if (!stopping) {
             setImmediate(do_spam, idx);
+          } else {
+            kill_spam_member();
           }
         });
       });
@@ -270,7 +332,7 @@ function do_spam(idx) {
       if (_tdg[idx]) {
         cookie = `_tdg=${_tdg[idx]}; ${cookie}; _TDG=${_tdg[idx]}`;
       }
-      let req1 = https.request(`https://${spam_target}${entry_path}`, {
+      let req1 = https.request(`https://${spam_targets_cache[idx].target_domain}${spam_targets_cache[idx].entry_path}`, {
         method: 'GET',
         headers: {
           'User-Agent': useragent,
@@ -310,6 +372,8 @@ function do_spam(idx) {
           // Proceed to next stage if not stopping.
           if (!stopping) {
             setImmediate(do_spam, idx);
+          } else {
+            kill_spam_member();
           }
         });
       });
@@ -343,11 +407,11 @@ function do_spam(idx) {
         cookie = `_tdg=${_tdg[idx]}; ${cookie}; _TDG=${_tdg[idx]}`;
       }
       let headers1 = {
-        'Host': spam_target,
+        'Host': spam_targets_cache[idx].target_domain,
         'User-Agent': useragent,
         'Connection': 'keep-alive',
-        'Origin': `https://${spam_target}`,
-        'Referer': `https://${spam_target}${phishing_page_path}`,
+        'Origin': `https://${spam_targets_cache[idx].target_domain}`,
+        'Referer': `https://${spam_targets_cache[idx].target_domain}${spam_targets_cache[idx].phishing_page_path}`,
         'Content-Length': `${content.length}`,
         'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8',
         'Accept': '*/*',
@@ -360,7 +424,7 @@ function do_spam(idx) {
         console.log(JSON.stringify(headers1, null, 2));
       }
 
-      let req1 = https.request(`https://${spam_target}${phishing_target_path}`, {
+      let req1 = https.request(`https://${spam_targets_cache[idx].target_domain}${spam_targets_cache[idx].phishing_target_path}`, {
         method: 'POST',
         headers: headers1
       });
@@ -375,20 +439,54 @@ function do_spam(idx) {
 
         resp1.on('end', function () {
           console.log('[%s:%s] Content -> %s (length: %d) ; Status Code -> %d', `0${idx}`.slice(-2), `000000${spam_sess_idx[idx]}`.slice(-7), credential, content.length, resp1.statusCode);
+          let resp_content = buf1.toString('utf8');
           if (debug_mode) {
-            console.log(buf1.toString('utf8'));
-          } else if (!stopping) {
-            // Received 429 Too Many Request, which means the C&C server has blocked this IP address.
-            if (resp1.statusCode == 429) {
-              stopping = true;
-              console.log('[%s:%s] ALERT: Received error code 429 (Too Many Requests), abort.', `0${idx}`.slice(-2), `000000${spam_sess_idx[idx]}`.slice(-7));
-              if (reli) {
-                // Close the interactive console or the process will halt and become an useless zombie.
-                reli.close();
-              }
-              return;
-            }
+            console.log(resp1.headers);
+            console.log(resp_content);
+          }
+          
+          // Received 429 Too Many Request, which means the C&C server has temporarily blocked this IP address.
+          if (resp1.statusCode === 429) {
+            if (sleep_when_429 > 0) {
+              console.log('[%s:%s] ALERT: Received error code 429 (Too Many Requests), mute %s for %s seconds.', `0${idx}`.slice(-2), `000000${spam_sess_idx[idx]}`.slice(-7), spam_targets_cache[idx].target_domain, sleep_when_429.toFixed(2));
 
+              // Put the server into hibernation mode.
+              if (target_pool.has(spam_targets_cache[idx].target_domain)) {
+                target_pool.delete(spam_targets_cache[idx].target_domain);
+                timeout_handles_targets[spam_targets_cache[idx].index] = setTimeout((server_entry) => {
+                  target_pool.set(server_entry.target_domain, server_entry);
+                }, sleep_when_429 * 1000, spam_targets_cache[idx]);
+              }
+            } else {
+              console.log('[%s:%s] ALERT: Received error code 429 (Too Many Requests), abort.', `0${idx}`.slice(-2), `000000${spam_sess_idx[idx]}`.slice(-7));
+
+              if (target_pool.has(spam_targets_cache[idx].target_domain)) {
+                target_pool.delete(spam_targets_cache[idx].target_domain);
+              }
+
+              if (target_pool.size <= 0) {
+                stopping = true;
+                kill_spam_member();
+                return;
+              }
+            }
+          } else if (resp1.statusCode === 200) {
+            let content_type = resp1.headers['content-type'] || resp1.headers['Content-Type'];
+            if (content_type === 'application/json') {
+              let json_content = null;
+              try {
+                json_content = JSON.parse(resp_content);
+              } catch (err) {
+                console.error('Server returned malformed JSON!');
+                console.error(err.stack);
+              }
+              if (json_content) {
+                console.log('[%s:%s] Message: %s %s', `0${idx}`.slice(-2), `000000${spam_sess_idx[idx]}`.slice(-7), json_content['message'], json_content['code'] === 0 ? '(GOOD)' : '(Unexpected)');
+              }
+            }
+          }
+          
+          if (!stopping) {
             // Reset cached data and stage
             lumen_sessions[idx] = null;
             _tdg[idx] = null;
@@ -396,6 +494,8 @@ function do_spam(idx) {
 
             // Loop
             setImmediate(do_spam, idx);
+          } else {
+            kill_spam_member();
           }
         });
       });
@@ -434,7 +534,15 @@ if (!debug_mode) {
     // If sonsole input is 'stop', flag the process to stop ASAP.
     if (line.toLowerCase() === 'stop') {
       stopping = true;
-      reli.close();
+
+      // Clear sleeping members;
+      for (let i = 0 ; i < parallel_count ; i++) {
+        if (timeout_handles_member[i]) {
+          clearTimeout(timeout_handles_member[i]);
+          timeout_handles_member[i] = null;
+          kill_spam_member();
+        }
+      }
     }
   });
 }