+++ changelog.md
@@ -1,8 +1,13 @@
 # Change Log
 
+## 1.3.0
+[+] Added 1 more server.
+[+] Added the ability to spam multiple phishing server in single process.
+[+] Added hibernation functionality that allows the process to run in background and revive after a certain time when receiving HTTP 429 error.
+
 ## 1.2.0
-[+] Add 3 more servers. The server in the previous version has been taken down successfully.  
-[+] Add extra cookie detection to bypass spam checks.  
+[+] Added 3 more servers. The server in the previous version has been taken down successfully.  
+[+] Added extra cookie detection to bypass spam checks.  
 [+] Major tweaks -- add more flexibilities for remote target paths.
   
 ## 1.1.0